combine use cases in packages

+++ app/src/main/java/com/itacit/healthcare/presentation/news/fragments/NewsDetailsFragment.java
@@ -15,8 +15,8 @@ import android.widget.TextView;
 
 import com.itacit.healthcare.R;
 import com.itacit.healthcare.data.network.interceptors.AuthInterceptor;
-import com.itacit.healthcare.domain.interactor.GetAuthorsUseCase;
-import com.itacit.healthcare.domain.interactor.GetNewsDetailsUseCase;
+import com.itacit.healthcare.domain.interactor.news.GetAuthorsUseCase;
+import com.itacit.healthcare.domain.interactor.news.GetNewsDetailsUseCase;
 import com.itacit.healthcare.presentation.base.fragments.BaseFragmentView;
 import com.itacit.healthcare.presentation.news.NewsActivity;
 import com.itacit.healthcare.presentation.news.mappers.AuthorMapper;
@@ -102,7 +102,7 @@ public class NewsDetailsFragment extends BaseFragmentView<NewsDetailsPresenter, 
     @Override
     protected NewsDetailsPresenter createPresenter() {
         long newsId = getArguments().getLong("newsId");
-        return new NewsDetailsPresenter(new GetNewsDetailsUseCase(newsId), new GetAuthorsUseCase(0, 10), new NewsDetailsMapper(), new AuthorMapper());
+        return new NewsDetailsPresenter(new GetNewsDetailsUseCase((int)newsId), new GetAuthorsUseCase(0, 10), new NewsDetailsMapper(), new AuthorMapper());
     }
 
     @Override
